test install pihole

+++ src/Installation/install-service.sh
@@ -41,8 +41,14 @@ bash /usr/local/osbox/project/sw-osbox-core/src/BashScripts/set_pihole_blocklist
 #bash /usr/local/osbox/project/sw-osbox-core/src/BashScripts/set_pihole_setupvars.sh
 sleep 5
 
-rm -rf /usr/local/osbox/project/sw-osbox-core/src/osbox-web
+
+if [ -d "/usr/local/osbox/project/sw-osbox-core/src/osbox-web" ]; then
+    rm -rf /usr/local/osbox/project/sw-osbox-core/src/osbox-web
+fi
+
+
 git clone https://github.com/jerryhopper/sw-osbox-core-web.git /usr/local/osbox/project/sw-osbox-core/src/osbox-web
+
 
 
 
@@ -68,6 +74,7 @@ if [ "$NETTYPE" == "STATIC" ];then
 else
    echo "finished,10,finished">/etc/osbox/setup.state
 fi
+
